SCIT Implements no rain zone correctly, calculates centroids
SCIT can now identify whether there is rain or not in a cluster (by averaging the intensity across the cell. sometime depending on the image makeup, no rain can show as a cluster, normally this is cluster 0, and was skipped over, but if there are multiple disconnected areas then they were recognised as separate clusters. This is now fixed to take the average, and if below threshold, identify as rain.
Also implements centroid caluclation of each cluster. This info is kept in objects kept in a list in the radar frame analysis variables.